Factor out the attachable framework from legacy.
