+++ lib/extensions/attachable.rb
@@ -1,0 +1,1 @@
+module Extensions::Attachable; end
