+++ lib/extensions/attachable/action_controller.rb
@@ -1,0 +1,1 @@
+module Extensions::Attachable::ActionController; end
