+++ lib/extensions/attachable/action_view.rb
@@ -1,0 +1,1 @@
+module Extensions::Attachable::ActionView; end
