+++ lib/extensions/attachable/action_view/helpers.rb
@@ -1,0 +1,1 @@
+module Extensions::Attachable::ActionView::Helpers; end
