+++ lib/extensions/attachable/active_record.rb
@@ -1,0 +1,1 @@
+module Extensions::Attachable::ActiveRecord; end
